Use more native methods of SmartProxy for validation
This uses more built in validators. It also creates a setting for the
public key file and derives it if not specified.

The only downside is that the validate_readable doesn't print a helpful
command to generate it if needed. Programmable settings can be used to
generate it on the fly if the parent directory is readable. Another
alternative is to introduce another custom validator.

It slightly abuses the dependency injection hook to configure the task
launcher registry.