Make dynflow an explicit dependency
Loading the smart_proxy_dynflow (or any Smart Proxy plugin really) in
load_classes does not work due to the way plugin initialization happens.
For this plugin to work, dynflow must be running. This uses requires to
validate that dynflow is actually correctly enabled and running. This
also means that if dynflow was loaded but ended up in a failed state
that REX SSH also ends up in a failed state.